Move anchor styles back to anchor directory
